requirements.txt
python-telegram-bot
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
schedule
google-generativeai
pandas